Chapter 5 results working again
The N-step model introduced a backwards incompatible change
which requires a Model to name a predictor class. This
swithches the Exception type raised to a named exception
which allows a Policy to provide the previous behavior.

Testing these changes yields the same results as those
written in Chapter 5 which gives high confidence this
is working as expected.