Fixing lambda, mu scaling bug in diff_eq plots
The lambda and mu values should not be scaled by the
timestep. This is already taken care of within the odeint
implementation.

We know this is correct because nearly the same results
are produced for timesteps of 1, 100, or 1000. The small
changes are due to numerical improvements with smaller
timesteps. In contrast, scaling lambda and mu produces
huge changes in the convergence rate.